Absorb most of core_config into dev_msgs
