+++ tt_metal/hw/inc/blackhole/core_config.h
@@ -4,50 +4,12 @@
 
 #pragma once
 
-#include <stdint.h>
+#include <cstdint>
 
-enum ProgrammableCoreType {
-    TENSIX     = 0,
-    ACTIVE_ETH = 1,
-    IDLE_ETH   = 2,
-    COUNT      = 3,
-};
-
-enum class AddressableCoreType : uint8_t {
-    TENSIX    = 0,
-    ETH       = 1,
-    PCIE      = 2,
-    DRAM      = 3,
-    HARVESTED = 4,
-    UNKNOWN   = 5,
-    COUNT     = 6,
-};
-
-enum class TensixProcessorTypes : uint8_t {
-    DM0    = 0,
-    DM1    = 1,
-    MATH0  = 2,
-    MATH1  = 3,
-    MATH2  = 4,
-    COUNT  = 5
-};
-
-enum class EthProcessorTypes : uint8_t {
-    DM0    = 0,
-    DM1    = 1,
-    COUNT  = 2
-};
-
-enum class DramProcessorTypes : uint8_t {
-    DM0     = 0,
-    COUNT   = 1
-};
-
-constexpr uint8_t MaxProcessorsPerCoreType = 5;
-constexpr uint8_t NumTensixDispatchClasses = 3;
-constexpr uint8_t NumEthDispatchClasses = 2;
-constexpr uint8_t NumDramDispatchClasses = 1;
-constexpr uint8_t noc_size_x = 17;
-constexpr uint8_t noc_size_y = 12;
+constexpr std::uint8_t NumTensixDispatchClasses = 3;
+constexpr std::uint8_t NumEthDispatchClasses = 2;
+constexpr std::uint8_t NumDramDispatchClasses = 1;
+constexpr std::uint8_t noc_size_x = 17;
+constexpr std::uint8_t noc_size_y = 12;
 #define ALLOCATOR_ALIGNMENT 64
 #define LOG_BASE_2_OF_ALLOCATOR_ALIGNMENT 6
